test for tipinfo
